Integrate segment_tool into item_tool
We're using common GTK controllers to handle events. This does not work well
when certain controllers act on the same input events. In our case item_tool
and segment_tool. I thought this could be fixed by giving segment_tool
preference. Change it's propagation level to TARGET. This raised another issue:
for line splitting the segment tool was used, but for line merging item_tool
was used. Hence, line segments are never merged.

Apart from that, the update handlers for both segment_tool and item_tool were identical.

To "enable" segments, the `gaphas.segment` module still needs to be included.
