+++ docs/segment.rst
@@ -7,12 +7,13 @@ The line segment functionality is an add-on, that will allow the user to add lin
    :align: center
 
 To use this behavior, import the ``gaphas.segment`` module and add ``LineSegmentPainter`` to the list of painters for the view.
-Add ``segment_tool`` to the view as a controller, in order to activate the split/merge segment behavior for ``Line`` instances.
+Splitting and merging of lines is supported by ``item_tool``, however
+to actually use it, the ``segment`` module needs to be imported.
 
     >>> from gaphas.view import GtkView
     >>> from gaphas.painter import PainterChain, ItemPainter, HandlePainter
     >>> from gaphas.tool import item_tool, scroll_tool, zoom_tool
-    >>> from gaphas.segment import LineSegmentPainter, segment_tool
+    >>> from gaphas.segment import LineSegmentPainter
     >>> view = GtkView()
     >>> view.painter = (
     ...     PainterChain()
@@ -20,7 +21,6 @@ Add ``segment_tool`` to the view as a controller, in order to activate the split
     ...     .append(HandlePainter(view))
     ...     .append(LineSegmentPainter(view.selection))
     ... )
-    >>> view.add_controller(segment_tool(view))
     >>> view.add_controller(item_tool(view))
     >>> view.add_controller(scroll_tool(view))
     >>> view.add_controller(zoom_tool(view))
