+++ gaphas/tool/itemtool.py
@@ -1,4 +1,5 @@
 import logging
+from functools import singledispatch
 from typing import Optional, Tuple, Union
 
 from gi.repository import Gdk, Gtk
@@ -6,7 +7,7 @@ from gi.repository import Gdk, Gtk
 from gaphas.aspect import HandleMove, Move, item_at_point
 from gaphas.canvas import ancestors
 from gaphas.connector import Handle
-from gaphas.geometry import distance_point_point_fast
+from gaphas.geometry import distance_line_point, distance_point_point_fast
 from gaphas.item import Item
 from gaphas.types import Pos
 from gaphas.view import GtkView
@@ -35,13 +36,14 @@ class DragState:
 
 def on_drag_begin(gesture, start_x, start_y, drag_state):
     view = gesture.get_widget()
+    pos = (start_x, start_y)
     selection = view.selection
     modifiers = (
         gesture.get_last_event(None).get_state()[1]
         if Gtk.get_major_version() == 3
         else gesture.get_current_event_state()
     )
-    item, handle = find_item_and_handle_at_point(view, (start_x, start_y))
+    item, handle = find_item_and_handle_at_point(view, pos)
 
     # Deselect all items unless CTRL or SHIFT is pressed
     # or the item is already selected.
@@ -63,6 +65,9 @@ def on_drag_begin(gesture, start_x, start_y, drag_state):
         selection.unselect_item(item)
         gesture.set_state(Gtk.EventSequenceState.DENIED)
         return
+
+    if not handle and item is view.selection.focused_item:
+        handle = maybe_split_segment(view, item, pos)
 
     selection.focused_item = item
     gesture.set_state(Gtk.EventSequenceState.CLAIMED)
@@ -113,6 +118,8 @@ def on_drag_end(gesture, offset_x, offset_y, drag_state):
     _, x, y = gesture.get_start_point()
     for moving in drag_state.moving:
         moving.stop_move((x + offset_x, y + offset_y))
+        if hasattr(moving, "item") and hasattr(moving, "handle"):
+            maybe_merge_segments(gesture.get_widget(), moving.item, moving.handle)
     drag_state.moving = set()
 
 
@@ -170,3 +177,60 @@ def handle_at_point(
         if h:
             return item, h
     return None, None
+
+
+@singledispatch
+class Segment:
+    def __init__(self, item, model):
+        raise TypeError
+
+    def split_segment(self, segment, count=2):
+        ...
+
+    def split(self, pos):
+        ...
+
+    def merge_segment(self, segment, count=2):
+        ...
+
+
+def maybe_split_segment(view, item, pos):
+    try:
+        segment = Segment(item, view.model)
+    except TypeError:
+        return None
+    else:
+        cpos = view.matrix.inverse().transform_point(*pos)
+        return segment.split(cpos)
+
+
+def maybe_merge_segments(view, item, handle):
+    handles = item.handles()
+
+    # don't merge using first or last handle
+    if handles[0] is handle or handles[-1] is handle:
+        return
+
+    # ensure at least three handles
+    handle_index = handles.index(handle)
+    segment = handle_index - 1
+
+    # cannot merge starting from last segment
+    if segment == len(item.ports()) - 1:
+        segment = -1
+    assert segment >= 0 and segment < len(item.ports()) - 1
+
+    before = handles[handle_index - 1]
+    after = handles[handle_index + 1]
+    d, p = distance_line_point(before.pos, after.pos, handle.pos)
+
+    if d > 4:
+        return
+
+    try:
+        Segment(item, view.model).merge_segment(segment)
+    except ValueError:
+        pass
+    else:
+        if handle:
+            view.model.request_update(item)
